Split the functions
