+++ core/node/interop-switch/src/chain/mod.rs
@@ -1,0 +1,5 @@
+mod destination;
+mod source;
+
+pub use destination::{DestinationChain, LocalDestinationChain};
+pub use source::SourceChain;
