+++ core/node/interop-switch/src/db/mod.rs
@@ -2,6 +2,7 @@ use zksync_basic_types::{L2ChainId, H256};
 
 use crate::types::{InteropBundle, InteropTrigger};
 pub(crate) mod in_memory;
+pub use in_memory::InMemoryDb;
 
 #[async_trait::async_trait]
 pub trait DbClient: Clone {
@@ -33,10 +34,4 @@ pub trait DbClient: Clone {
     ) -> anyhow::Result<()>;
 
     async fn get_last_processed_block(&mut self, src_chain_id: L2ChainId) -> anyhow::Result<u64>;
-
-    async fn inject_new_fee_bundle(
-        &mut self,
-        tx_hash: H256,
-        fee_bundle: Vec<u8>,
-    ) -> anyhow::Result<()>;
 }
